cmd: move `ProvisionerServer` implementation into `internal/`
This code is not really part of the application entry-point, hence
moving it into its own package.
